al fin es viernes

+++ app/GraphQL/Mutations/TecnicoHabilidadMutations.php
@@ -10,6 +10,7 @@ class TecnicoHabilidadMutations{
     public function assign($root,array $args){
         #return $tecnico=Tecnico_Habilidad::create($args);
         $id = $args['technicianId'];
+        //dd($id);
         $envHabilidades = $args['skills'];
         $habilidades=[];
         foreach ($envHabilidades as $recHabilidad) {
@@ -22,19 +23,26 @@ class TecnicoHabilidadMutations{
 
             $habilidades[] = $habilidadTecnico;
         }
-        return $habilidades;
-        /*return [
-            'message' => 'habilidades asignadas al tecnico OK'
-            'assings' => $habilidades
-        ];*/
+
+        $technician = Tecnico::find($id);
+        //return $habilidades;
+        return [
+            'message' => 'habilidades asignadas al tecnico OK' ,
+            'technical' => $technician,
+            'skills' => $habilidades
+        ];
     }
     public function update($root,array $args){
         $tecnicoId = $args['id'];
         $habilidades = $args['skills'];
-
+        $technician = Tecnico::find($tecnicoId);
+        if(!$technician){
+            return [
+                'message' => 'No existe tecnico'
+            ];
+        }
         // Eliminar las habilidades existentes del técnico
         Tecnico_Habilidad::where('technicianId', $tecnicoId)->delete();
-
         // Guardar las nuevas habilidades
         foreach ($habilidades as $habilidad) {
             Tecnico_Habilidad::create([
@@ -45,8 +53,12 @@ class TecnicoHabilidadMutations{
             ]);
         }
         $habilidades = Tecnico_Habilidad::where('technicianId', $tecnicoId)->get();
-        // Retornar las nuevas habilidades
-        return $habilidades;
+        //return $habilidades;
+        return [
+            'message' => 'habilidades actualizadas al tecnico OK' ,
+            'technical' => $technician,
+            'skills' => $habilidades
+        ];
     }
 
 }
